Add setup.py file, rearrange repository structure
A setuptools-based setup.py file is now included in the repository,
in order to support easier installation. ("python setup.py install",
for example). To accomodate this the sources have been moved inside
a "quasselgrep" directory; this will be installed as a Python package.

I added a "main" function to the quasselgrep script (and renamed it
to quasselgrep/quasselgrep.py) that's used as an entry point for
setuptools to auto-generate a "quasselgrep" script that will be installed
into the system-wide scripts directory.

Because current PyPI cannot parse Markdown long_descriptions, and because
the de facto standard long_description is just the contents of the README
file read into setuptools, I modified the standard setup.py script
slightly by making it first attempt to use pypandoc to convert the README
from MD to rST. If this fails due to an ImportError, the markdown
README is used unmodified as the long_description; this means that
pypandoc is not a dependency for end-users, but if you want to upload
this package to PyPI, if you have pypandoc installed on your own system
you can use it to ensure the description is nicely formatted.

quasselgrep appears to still run after these changes, but I haven't
gotten around to setting it up myself so that should be verified.

(This is slightly biased to how I myself set up Python projects, but
I based the setup.py file on this one:
https://pypi.python.org/pypi/python-project-template/0.1.4.).